requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.2
groq>=0.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.0.0
semantic-router==0.0.36
